requirements.txt
ccxt>=4.1.13
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
aiohttp==3.8.6
python-binance>=1.0.16
websocket-client>=1.3.1

pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0

python-telegram-bot>=20.7
schedule>=1.2.0

psutil>=5.9.0
colorama>=0.4.6

jupyter>=1.0.0
ipython>=8.0.0 